Init Commit
-I input 2 basic autonomous and teleop files from previous years that
should give a solid basis for the competition,
-The autonomous features 2 potential driving methods along with a few
helper methods: one based on the phone's internal time keeping system as
the match goes on and one based on the traditional "wait" system that
stalls actions in the loop function.
-Teleop features a basec 4-wheel drive system along with a telemetry
statement to ease testing. In between the two there is also a commented
out holonomic drive train formula that will work for sideways strafing
should we need to use that in this competition...